게시글 delete post

+++ articles/views.py
@@ -45,5 +45,8 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+
+    if request.method == 'POST':
+        article.delete()
+        
     return redirect('articles:dinner')
